+++ CHANGELOG.md
@@ -3,6 +3,16 @@ This file documents all changes to Argus-frontend. This file is primarily meant 
 
 ## [Unreleased]
 
+### Added
+- Handling of both network errors, and errors with fetching existing user in _Login view_ .
+
+
+### Fixed
+- Bug with misleading error message in Login view on connection problems.
+
+
+
+
 ## [v1.9.0] - 2023-03-17
 
 ### Changed